docs: partial fix for planned elements
