+++ docs/_includes/partials/component/header.njk
@@ -1,21 +1,10 @@
+{%- set tagName = tagName or doc.tagName or element.tagName or doc.docsPage.tagName -%}
 {%- set headingId = (heading if heading else title) | lower | slug -%}
-{%- set headingContent = heading or title -%}
-{%- set manifest = doc and doc.docsPage and doc.docsPage.manifest -%}
-{%- if subnav.collection -%}
-  {% set tabs = collections[subnav.collection] | sort(attribute="data.subnav.order") %}
-{%- endif -%}
-{%- if manifest -%}
-  {% set tabs = doc.tabs %}
-  {% set demos = manifest.getDemos(doc.docsPage.tagName) %}
-  {% set planned = doc.docsPage.tagName | isPlanned %}
-  {% set prettyName -%}{{ (doc.alias or doc.slug or element.slug) | deslugify | capitalize }}{% endset %}
-  {% set headingContent -%}
-    {{ prettyName }}
-    {% if planned %}
-    <rh-tag icon="notification-fill" color="gray">Planned</rh-tag>
-    {% endif %}
-  {% endset %}
-{%- endif -%}
+{%- set manifest = (doc and doc.docsPage and doc.docsPage.manifest) -%}
+{%- if manifest == (doc and doc.docsPage) -%}{% set manifest = null %}{% endif %}
+{% set prettyName -%}{{ (doc.alias or doc.slug or element.slug or '') | deslugify | capitalize }}{% endset %}
+{% set tabs = doc.tabs or (collections[subnav.collection or ''] | sort(attribute="data.subnav.order")) %}
+{% set demos = manifest and manifest.getDemos(doc.docsPage.tagName) %}
 
 <script type="module" data-helmet>
   import '/assets/javascript/elements/uxdot-header.js';
@@ -25,12 +14,16 @@
 <script type="module" data-helmet>
   import '@rhds/elements/rh-subnav/rh-subnav.js';
 </script>
-
 {% endif %}
 
 <uxdot-header {% if tabs %}has-subnav{% endif %} {% if tokenSearch %}has-search {% endif %}>
   {#- TODO: cheap hack here to fix data spaghetti from demos page -#}
-  <h1 id="{{ headingId }}" class="page-title">{{ headingContent | replace('Demos |', '') }}</h1>
+  <h1 id="{{ headingId }}" class="page-title">
+    {{ (prettyName or heading or title) | replace('Demos |', '') }}
+    {% if doc.planned %}
+    <rh-tag icon="notification-fill" color="gray">Planned</rh-tag>
+    {% endif %}
+  </h1>
   {% if tokenSearch %}{% include 'partials/component/token-search.html' %}{% endif %}
   {% if tabs %}
   <rh-subnav slot="subnav" accessible-label="{{ heading }}">
@@ -39,7 +32,7 @@
     {%- set pageUrl = (doc and doc.permalink | replace('index.html', '')) or page.url -%}
     {%- set title = tab.data.title or tab.pageTitle -%}
     {%- set active = href == pageUrl -%}
-    <a data-page-url="{{ page.url }}" href="{{ href }}" {{ 'active' if active }}>{{ title | capitalize }}</a>
+    <a href="{{ href }}" {{ 'active' if active }} {{ 'hidden' if hidden }}>{{ title | capitalize }}</a>
   {%- endfor -%}
   </rh-subnav>
   {% endif %}
