+++ docs/_plugins/element-docs.cjs
@@ -2,10 +2,11 @@
 
 // @ts-check
 const path = require('node:path');
-const { glob } = require('node:fs/promises');
+const { glob, readFile } = require('node:fs/promises');
 const _slugify = require('slugify');
 const slugify = typeof _slugify === 'function' ? _slugify : _slugify.default;
 const capitalize = require('capitalize');
+const yaml = require('js-yaml');
 
 /** @param {string} path */
 const exists = async path => {
@@ -17,6 +18,10 @@ const exists = async path => {
     return false;
   }
 };
+
+async function readYaml(filePath) {
+  return yaml.load(await readFile(path.join(cwd, filePath), 'utf8'));
+}
 
 const cwd = process.cwd();
 
@@ -46,6 +51,22 @@ function alphabeticallyBySlug(a, b) {
 
 /** @param {import('@11ty/eleventy').UserConfig} eleventyConfig */
 module.exports = function(eleventyConfig) {
+  async function isPlanned(tagName) {
+    const { getAllManifests } =
+      await import('@patternfly/pfe-tools/custom-elements-manifest/custom-elements-manifest.js');
+    for (const manifest of getAllManifests()) {
+      if (manifest.declarations.has(tagName)) {
+        return false;
+      }
+    }
+    const repoStatus =
+      eleventyConfig.globalData?.repoStatus ?? await readYaml('./docs/_data/repoStatus.yaml');
+    const element = repoStatus.find(element => element.tagName === tagName);
+    return element?.libraries.find(library => library.name === 'RH Elements')?.status === 'Planned';
+  }
+
+  eleventyConfig.addFilter('isPlanned', isPlanned);
+
   eleventyConfig.addCollection('elementDocs', async function() {
     const { pfeconfig } = eleventyConfig?.globalData ?? {};
     /**
@@ -82,6 +103,7 @@ module.exports = function(eleventyConfig) {
       /** urls for related links */
       return {
         tagName,
+        planned: await isPlanned(tagName),
         filePath,
         fileExists,
         hasLightdom,
@@ -102,15 +124,14 @@ module.exports = function(eleventyConfig) {
 
     try {
       const { DocsPage } = await import('@patternfly/pfe-tools/11ty/DocsPage.js');
-      const {
-        getAllManifests,
-      } = await import(
-        '@patternfly/pfe-tools/custom-elements-manifest/custom-elements-manifest.js'
-      );
+      const { getAllManifests } =
+        await import('@patternfly/pfe-tools/custom-elements-manifest/custom-elements-manifest.js');
 
       const customElementsManifestDocsPages = await eleventyConfig.globalData?.elements();
+
       const filePaths = (await Array.fromAsync(glob(`elements/*/docs/*.md`, { cwd })))
           .filter(x => x.match(/\d{1,3}-[\w-]+\.md$/)); // only include new style docs
+
       const elementDocFilePaths =
         Array.from(new Set([
           ...filePaths,
@@ -119,8 +140,7 @@ module.exports = function(eleventyConfig) {
           // (elementDocs) should be refactored, and the elements/*/docs/*.md files should be used
           // only for markdown content, but the code and demos tabs should be fully generated, with
           // the XX-code.md content interjected, if any.
-          ...await Array.fromAsync(glob('elements/*', { cwd }), x => `${x}/docs/30-code.md`
-          ),
+          ...await Array.fromAsync(glob('elements/*', { cwd }), x => `${x}/docs/30-code.md`),
         ]));
 
       const elementDocs = await Promise.all(elementDocFilePaths.map(async filePath => {
@@ -131,13 +151,16 @@ module.exports = function(eleventyConfig) {
             ?? new DocsPage(manifest);
 
         const demosUrl = `/elements/${props.slug}/demos/`;
-        const tabs = await Promise.all(elementDocFilePaths
+        let tabs = await Promise.all(elementDocFilePaths
             .filter(x => x.split('/docs/').at(0) === (`elements/${props.tagName}`))
             .sort()
         // todo: avoid calling getProps twice, it's expensive
             .map(x => getProps(x)));
         // Add 'virtual' demos page, it's generated by demos.11ty.cjs
         tabs.splice(-1, 0, { ...props, pageTitle: 'Demos', href: demosUrl });
+        if (props.planned) {
+          tabs = tabs.filter(x => x.pageSlug !== 'demos' && x.pageSlug !== 'code');
+        }
         return { docsPage, tabs, ...props };
       }));
       elementDocs.sort(alphabeticallyBySlug);
